Integrated recommendActivities into getWeather, completed response with status and json; refactored cityLookup
