+++ api/utils/cityLookup.js
@@ -1,0 +1,24 @@
+const cityLookup = {
+  'brighton': {
+    "lat": 50.8214626,
+    "lon": -0.1400561,
+  },
+  'lisbon': {
+    "lat": 38.7077507,
+    "lon": -9.1365919
+  },
+  'valencia': {
+    "lat": 39.4697065,
+    "lon": -0.3763353
+  },
+  'riga': {
+    "lat": 56.9493977,
+    "lon": 24.1051846
+  },
+  'santorini': {
+    "lat": 36.4622122,
+    "lon": 25.3757257
+  }
+}
+
+module.exports = cityLookup;
